remove dayz_utils file multi-table
